requirements.txt
faiss-cpu
numpy
openai
pypdf
sentence-transformers
transformers
torch
accelerate
scikit-learn
